boot laravel in backend
When a page of the add-on is opened in the Redaxo backend, the pages/index.php is called. So we basically do everything that normally would happen in public/index.php. Plus a bit of cleanup that is necessary because Redaxo makes some changes that influence the generation of urls.